worker/requirements.txt
django==1.11.*
djangorestframework==3.7.*
structlog==18.*
requests==2.18.*
boto3==1.5.*
